Finished working on new additional enemy
Finished the working on the flying demon enemy for the game should work as expected now it can hurt the player and will turn depending on the player position relative to the enemy. Also now the enemy can be hurt and killed I've given the enemy a two shot lifespan so far and upon the completion of two shots the enemy will explode aka death. In the progress of adding another enemy called Hell Beast who should be completed hopefully by tomorrow which should be my last push for the enemies branch